+++ crm/views/contract.py
@@ -3,19 +3,19 @@ from typing_extensions import Annotated
 from crm.models.contract import Contract
 from crm.models.client import Client
 from datetime import datetime
+from crm.auth import auth_required, check_user_and_permissions
 
 
 app = typer.Typer()
 
 
 def find_client():
-    first_name = typer.prompt("Enter first_name of the client")
-    last_name = typer.prompt("Enter last_name of the client")
+    client_id = typer.prompt("Enter the id of the client")
     try:
-        client = Client.get(first_name=first_name, last_name=last_name)
+        client = Client.get(id=client_id)
     except Client.DoesNotExist:
         typer.echo("Client non trouvé")
-        find_client()
+        return
     return client
 
 
@@ -27,6 +27,7 @@ def sign(contract):
 
 
 @app.command()
+@auth_required
 def create_contract(
     total_amount: int = typer.Option(
         ..., "-t", prompt=True, help="Total amount of the contract"
@@ -38,15 +39,21 @@ def create_contract(
         ..., "--sign", prompt=True, help="Indicate if the contract is signed or not"
     ),
     created_at: datetime = datetime.now(),
+    user=None,
 ):
     try:
+        if not check_user_and_permissions(user, "create-contract"):
+            return
         client = find_client()
+
+        commercial_contact = client.epic_events_contact
         contract = Contract.create(
             client=client,
             total_amount=total_amount,
             remaining_amount=remaining_amount,
             status=status,
             created_at=created_at,
+            commercial_contact=commercial_contact,
         )
         typer.echo("Contract created successfully")
         typer.echo(f"Contract ID: {contract.id}")
@@ -55,11 +62,17 @@ def create_contract(
 
 
 @app.command()
+@auth_required
 def delete_contract(
-    contract_id: int = typer.Option(..., "-i", prompt=True, help="id of the contract")
+    contract_id: int = typer.Option(..., "-i", prompt=True, help="id of the contract"),
+    user=None,
 ):
     try:
         contract = Contract.get(id=contract_id)
+        if not check_user_and_permissions(
+            user, "delete-contract"
+        ) and not user.has_permission_own(contract.commercial_contact):
+            return
         contract.delete_instance()
         typer.echo(f"contract with id: {contract_id} deleted succesfuly")
 
@@ -70,8 +83,11 @@ def delete_contract(
 
 
 @app.command(name="list-contracts")
-def list_contracts():
+@auth_required
+def list_contracts(user=None):
     try:
+        if not check_user_and_permissions(user, "list-contract"):
+            return
         contracts = Contract.select()
         if not contracts:
             typer.echo("No elements found")
@@ -88,15 +104,19 @@ def list_contracts():
 
 
 @app.command()
+@auth_required
 def get_contract(
     contract_id: int = typer.Option(
         ...,
         "-i",
         prompt="Enter id of the contract you want to see details about",
         help="id of the contract",
-    )
+    ),
+    user=None,
 ):
     try:
+        if not check_user_and_permissions(user, "get-contract"):
+            return
         contract = Contract.get(id=contract_id)
         status = sign(contract)
         typer.echo(
@@ -111,13 +131,19 @@ def get_contract(
 
 
 @app.command()
+@auth_required
 def update_contract(
     contract_id: Annotated[
         int, typer.Option(..., prompt=True, help="ID of the contract")
     ],
+    user=None,
 ):
     try:
         contract = Contract.get(id=contract_id)
+        if not check_user_and_permissions(
+            user, "update-contract"
+        ) and not user.has_permission_own(contract.commercial_contact):
+            return
         status = sign(contract)
         typer.echo(
             "Details of the contract: "
